Move PublishingPanel to wagtail.admin.panels.publishing_panel
